fixed false warning for high k

+++ src/mash/CommandDistance.cpp
@@ -125,8 +125,8 @@ int CommandDistance::run() const
     
     Sketch sketch;
     
-    int lengthThreshold = (parameters.warning * pow(parameters.protein ? 20 : 4, parameters.kmerSize)) / (1. - parameters.warning);
-    int lengthMax;
+    uint64_t lengthThreshold = (parameters.warning * pow(parameters.protein ? 20 : 4, parameters.kmerSize)) / (1. - parameters.warning);
+    uint64_t lengthMax;
     double randomChance;
     int kMin;
     string lengthMaxName;
